refactor(teams): one preset resolver, one hostPort, no dead nameSlots (STORE-4)
TeamRosterResolution.nameSlots had no production reader; resolveAgentKeys duplicated TeamPlan's resolvePresetAgents, which now lives in TeamRoster (TeamPlan already imports it, so no cycle). Two private hostPort copies and one inline identity-catch tryPromise now use @common/hostPort.
